Add `CollectorConfig` base class for source collector configs (#110)
This PR adds the class
`zabbix_auto_config.sourcecollector.CollectorConfig`, which is a
Pydantic model that can be used as a base class for source collector
config validation. The class provides the
`CollectorConfig.from_kwargs()` method, which constructs a config from
kwargs passed into a `collect()` function.

With the functionality added in this PR, source collectors can do away
with dict key access of collector config values in `collect()`, and
instead define Pydantic models to use for their configs.